move SpawnerTraceExt back into Spawner
